requirements.txt
pandas>=1.5
numpy>=1.23
requests>=2.28
scipy>=1.9
scikit-learn>=1.2
matplotlib>=3.6
seaborn>=0.12
openpyxl>=3.0
pytest>=7.0
python-dotenv>=0.21

